Simplify and improve MetadataReader tasks and interop with Futures

+++ src/audio/tagging/MetadataReader.java
@@ -3,10 +3,8 @@ package audio.tagging;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.concurrent.ExecutionException;
 import java.util.function.BiConsumer;
 
-import javafx.application.Platform;
 import javafx.concurrent.Task;
 import javafx.scene.media.Media;
 
@@ -20,24 +18,62 @@ import util.file.AudioFileFormat.Use;
 
 import static services.database.Db.em;
 import static util.async.Async.runFX;
-import static util.async.Async.runNew;
-import static util.dev.Util.log;
-import static util.dev.Util.noØ;
+import static util.dev.Util.*;
+import static util.functional.Util.stream;
 
 /**
- * This class plays the role of static factory for Metadata. It can read files
- * and construct Metadata items. The class makes use of concurrency.
- *
- * ------------- I/O OPERATION ---------------
- * Everything is read once at object creation time. There is no additional READ
- * operation after Metadata has been created ( with the exception of some of
- * the non-tag (possibly external) information - cover, etc)
  *
  * @author Martin Polakovic
  */
 public class MetadataReader{
 
-    private static Task<List<Metadata>> buildReadMetadata(Collection<? extends Item> items, BiConsumer<Boolean, List<Metadata>> onEnd){
+	// TODO: return Try
+	/**
+	 * Reads metadata for specified item.
+	 * <p/>
+	 * Involves I/O operation and blocks thread. Throws exception if executed on fx application thread.
+	 *
+	 * @param item item to read metadata for
+	 * @return metadata for specified item or {@link Metadata#EMPTY} if any error. Never null.
+	 */
+	public static Metadata readMetadata(Item item){
+		throwIfFxThread();
+
+		if (item.isCorrupt(Use.APP)) {
+			return Metadata.EMPTY;
+		}
+
+		if (item.isFileBased()) {
+			AudioFile f = MetaItem.readAudioFile(item.getFile());
+			return f == null ? Metadata.EMPTY : new Metadata(f);
+		} else {
+			// TODO: implement properly
+			try {
+				Media m = new Media(item.getURI().toString());
+				//            m.getMetadata().forEach((String s, Object o) -> {
+				//                System.out.println(s + " " + o);
+				//            });
+
+				// make a playlistItem and covert to metadata //why? // not 100%sure...
+				// because PlaylistItem has advanced update() method? // probably
+				return new PlaylistItem(item.getURI(), "", "", m.getDuration().toMillis()).toMeta();
+			} catch (IllegalArgumentException | UnsupportedOperationException e){
+				log(MetadataReader.class).error("Error creating metadata for non file based item: {}",item);
+				return item.toMeta();
+			}
+		}
+	}
+
+	/**
+	 * Creates task that reads metadata for specified items.
+	 *
+	 * @param items list of items to read mretadata for
+	 * @param onEnd procedure to execute upon finishing this task providig the result and success flag. Must not be null.
+	 * @return the task reading the files returning all successfully read metadata
+	 *
+	 * @throws NullPointerException if any parameter null
+	 */
+	public static Task<List<Metadata>> buildReadMetadataTask(Collection<? extends Item> items, BiConsumer<Boolean, List<Metadata>> onEnd) {
 	    noØ(items);
 	    noØ(onEnd);
         return new SuccessTask<List<Metadata>,SuccessTask>("Reading metadata", onEnd){
@@ -56,8 +92,7 @@ public class MetadataReader{
                         return metadatas;
                     }
 
-                    // create metadata
-                    Metadata m = create(item);
+                    Metadata m = readMetadata(item);
                     // on fail
                     if (m.isEmpty()) skipped++;
                     // on success
@@ -73,156 +108,6 @@ public class MetadataReader{
             }
         };
     }
-
-    /**
-     * Creates list of Metadata for provided items. Use to read multiple files
-     * at once. The work runs on background thread. The procedures executed on
-     * task completion will be automatically executed from FXApplication thread.
-     * <p/>
-     * This method returns {@link Task} doing the work, which allows binding to
-     * its properties (for example progress) and more.
-     * <p/>
-     * When any error occurs during the reading process, the reading will stop
-     * and return all obtained metadata.
-     * <p/>
-     * The result of the task is list of metadatas (The list will not be null
-     * nor contain null values) if task finshes sccessfully or null otherwise.
-     * <p/>
-     * Calling this method will immediately start the reading process (on
-     * another thread).
-     *
-     * @param items List of items to read.
-     * @param onEnd procedure to execute upon finishing this task providig
-     * the result and success flag.
-     * Must not be null.
-     * @return task reading the files returning item's metadatas on successful
-     * completion or all successfully obtained metadata when any error occurs.
-     *
-     * @throws NullPointerException if any parameter null
-     */
-    public static Task<List<Metadata>> readMetadata(Collection<? extends Item> items, BiConsumer<Boolean, List<Metadata>> onEnd){
-        // create task
-        final Task task = buildReadMetadata(items, onEnd);
-
-        // run immediately and return task
-        runNew(task);
-        return task;
-    }
-
-    /**
-     * Transforms items into their metadatas by reading the files.
-     * For asynchronouse use only.
-     * Items for which reading fails are ignored.
-     */
-    public static List<Metadata> readMetadata(Collection<? extends Item> items){
-        List<Metadata> metadatas = new ArrayList<>();
-
-        for (Item item: items){
-            // create metadata
-            Metadata m = create(item);
-            if (!m.isEmpty()) metadatas.add(m);
-        }
-
-        return metadatas;
-    }
-
-    /**
-     * Reads {@link Metadata} for specified item.
-     * When error occurs during reading {@link Metadata#EMPTY} will be
-     * returned.
-     * <p/>
-     * Incurs costly I/O.
-     * Avoid using this method in loops or in chains on main application thread.
-     *
-     * @param item
-     * @return metadata for specified item or {@link Metadata#EMPTY} if error.
-     * Never null.
-     */
-    public static Metadata create(Item item){
-//        // handle corrupt item
-        if (item.isCorrupt(Use.APP)){
-            return Metadata.EMPTY;        // is this good way to handle corrupt item? //no.
-        }
-        // handle items with no file representation
-        if (!item.isFileBased()){
-            return createNonFileBased(item);
-        }
-        // handle normal item
-        else {
-            afile = MetaItem.readAudioFile(item.getFile());
-            return (afile == null) ? item.toMeta() : new Metadata(afile);
-        }
-    }
-
-    private static AudioFile afile;
-
-    /**
-     * Reads {@link Metadata} for specified item. Runs on background thread.
-     * Calling this method will immediately start the execution. The procedures
-     * executed on task completion will always be executed from FXApplication
-     * thread.
-     * <p/>
-     * This method returns {@link Task} doing the work, which allows binding to
-     * its properties (for example progress) and more.
-     * <p/>
-     * The result of the task is nonempty Metadata if task finshes successfully
-     * or null otherwise.
-     *
-     * @param item item to read metadata for. Must not be null.
-     * @param onFinish procedure to execute upon finishing this task providig
-     * the result and success flag.
-     * Must not be null.
-     * @return task reading the file returning its metadata on successful task
-     * completion or nothing when any error occurs. Never null.
-     * @throws NullPointerException if any parameter null
-     */
-    public static Task<Metadata> create(Item item, BiConsumer<Boolean, Metadata> onFinish){
-	    noØ(item);
-	    noØ(onFinish);
-
-        Task<Metadata> task = new Task<>(){
-            @Override protected Metadata call() throws Exception {
-                return create(item);
-            }
-        };
-        task.setOnSucceeded( e ->
-            Platform.runLater(() -> {
-                try {
-                    onFinish.accept(true, task.get());
-                } catch (InterruptedException | ExecutionException ex){
-                    log(MetadataReader.class).error("Reading metadata failed for: {}",item);
-                    onFinish.accept(false, null);
-                }
-            })
-        );
-        task.setOnFailed( e ->
-            Platform.runLater(() -> {
-                log(MetadataReader.class).error("Reading metadata failed for: {}", item);
-                onFinish.accept(false, null);
-            })
-        );
-
-        // run immediately and return task
-        runNew(task);
-        return task;
-    }
-
-    static private Metadata createNonFileBased(Item item){
-        try {
-            Media m = new Media(item.getURI().toString());
-//            m.getMetadata().forEach((String s, Object o) -> {
-//                System.out.println(s + " " + o);
-//            });
-
-            // make a playlistItem and covert to metadata //why? // not 100%sure...
-            // because PlaylistItem has advanced update() method? // probably
-            return new PlaylistItem(item.getURI(), "", "", m.getDuration().toMillis()).toMeta();
-        } catch (IllegalArgumentException | UnsupportedOperationException e){
-            log(MetadataReader.class).error("Error creating metadata for non file based item: {}",item);
-            return item.toMeta();
-        }
-    }
-
 
     /**
      * Creates a task that:
@@ -256,7 +141,7 @@ public class MetadataReader{
                         m = em.find(Metadata.class, Metadata.metadataID(item.getURI()));
                         if (m == null) {
                             MetadataWriter.useNoRefresh(item, MetadataWriter::setLibraryAddedNowIfEmpty);
-                            m = create(item);
+                            m = readMetadata(item);
 
                             if (m.isEmpty()) {
 	                            skipped.add(item);
@@ -280,8 +165,6 @@ public class MetadataReader{
                 }
 
                 em.getTransaction().commit();
-
-	            // update library model
                 runFX(Db::updateInMemoryDBfromPersisted);
 
                 // update progress
@@ -294,44 +177,53 @@ public class MetadataReader{
         };
     }
 
-    public static Task<Void> removeMissingFromLibrary(BiConsumer<Boolean,Void> onEnd){
-        // create task
-        final Task<Void> task = new SuccessTask<>("Removing missing items from library",onEnd){
+	/**
+	 * @return a task that removes from library all items, which refer to non-existent files
+	 */
+	public static Task<Void> buildRemoveMissingFromLibTask() {
+        return new Task<>() {
+		    private final StringBuffer sb = new StringBuffer(40);
             private int all = 0;
             private int completed = 0;
             private int removed = 0;
 
+            {
+	        	updateTitle("Removing missing items from library");
+	        }
+
             @Override
-            protected Void call() throws Exception {                    //long timeStart = System.currentTimeMillis();
-                List<Metadata> library_items = Db.getAllItems();
-                all = library_items.size();
+            protected Void call() throws Exception {
+                List<Metadata> libraryItems = stream(Db.getAllItems()).filter(m -> !m.isEmpty() && m.isFileBased()).toList();
+                all = libraryItems.size();
+
                 em.getTransaction().begin();
 
-                for (Metadata m : library_items){
-                    completed++;
+                for (Metadata m : libraryItems) {
                     if (isCancelled()) break;
 
-                    if (!m.getFile().exists()) {
+                    completed++;
+
+                    if (m.isFileBased() && !m.getFile().exists()) {
                         em.remove(m);
                         removed++;
                     }
+
+	                // update state
                     updateMessage(all,completed,removed);
                     updateProgress(completed, all);
                 }
 
                 em.getTransaction().commit();
-                // update library model
                 runFX(Db::updateInMemoryDBfromPersisted);
 
                 // update state
                 updateMessage(all,completed,removed);
-                updateProgress(completed, all);                     //System.out.println((System.currentTimeMillis()-timeStart));
+                updateProgress(completed, all);
 
                 return null;
             }
 
-            @Override
-            protected void updateMessage(int all, int done, int removed) {
+            private void updateMessage(int all, int done, int removed) {
                 sb.setLength(0);
                 sb.append("Completed ");
                 sb.append(all);
@@ -344,9 +236,5 @@ public class MetadataReader{
             }
 
         };
-
-        // run immediately and return task
-        runNew(task);
-        return task;
     }
 }
